requirements.txt
# OSDU Record Viewer Requirements
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0

# Optional: For production deployment
gunicorn==21.2.0

# Optional: For development
python-dotenv[cli]==1.0.0